grms/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pg8000==1.31.2
pydantic-settings==2.7.0
passlib==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
